// frontend-master/frontend-master/src/services/enums/signature.ts
export enum CADESCOM_CADES_TYPE {
  CADESCOM_CADES_BES = 1,
  CADESCOM_CADES_DEFAULT = 0,
  CADESCOM_CADES_T = 5,
  CADESCOM_CADES_X_LONG_TYPE_1 = 93,
}

export enum CADESCOM_CONTENT_ENCODING_TYPE {
  CADESCOM_BASE64_TO_BINARY = 1,
  CADESCOM_STRING_TO_UCS2LE = 0,
}
// ...
export enum CADESCOM_HASH_ALGORITHM {
  CADESCOM_HASH_ALGORITHM_SHA1 = 0,
  CADESCOM_HASH_ALGORITHM_MD2 = 1,
  CADESCOM_HASH_ALGORITHM_MD4 = 2,
  CADESCOM_HASH_ALGORITHM_MD5 = 3,
  CADESCOM_HASH_ALGORITHM_SHA_256 = 4,
  CADESCOM_HASH_ALGORITHM_SHA_384 = 5,
  CADESCOM_HASH_ALGORITHM_SHA_512 = 6,
  CADESCOM_HASH_ALGORITHM_CP_GOST_3411 = 100,
  CADESCOM_HASH_ALGORITHM_CP_GOST_3411_2012_256 = 101,
  CADESCOM_HASH_ALGORITHM_CP_GOST_3411_2012_512 = 102,
}
// ...
export enum CAPICOM_STORE_LOCATION {
  CAPICOM_MEMORY_STORE = 0,
  CAPICOM_LOCAL_MACHINE_STORE = 1,
  CAPICOM_CURRENT_USER_STORE = 2,
  CAPICOM_ACTIVE_DIRECTORY_USER_STORE = 3,
  CAPICOM_SMART_CARD_USER_STORE = 4,
}

export enum CAPICOM_STORE_NAME {
  CAPICOM_MY_STORE = 'My',
  CAPICOM_CA_STORE = 'Ca',
  CAPICOM_OTHER_STORE = 'AddressBook',
  CAPICOM_ROOT_STORE = 'Root',
}

export enum CAPICOM_STORE_OPEN_MODE {
  CAPICOM_STORE_OPEN_READ_ONLY = 0,
  CAPICOM_STORE_OPEN_READ_WRITE = 1,
  CAPICOM_STORE_OPEN_MAXIMUM_ALLOWED = 2,
  CAPICOM_STORE_OPEN_EXISTING_ONLY = 128,
  CAPICOM_STORE_OPEN_INCLUDE_ARCHIVED = 256,
}

export enum CAPICOM_CERTIFICATE_FIND_TYPE {
  CAPICOM_CERTIFICATE_FIND_SHA1_HASH = 0,
  CAPICOM_CERTIFICATE_FIND_SUBJECT_NAME = 1,
  CAPICOM_CERTIFICATE_FIND_ISSUER_NAME = 2,
  CAPICOM_CERTIFICATE_FIND_ROOT_NAME = 3,
  CAPICOM_CERTIFICATE_FIND_TEMPLATE_NAME = 4,
  CAPICOM_CERTIFICATE_FIND_EXTENSION = 5,
  CAPICOM_CERTIFICATE_FIND_EXTENDED_PROPERTY = 6,
  CAPICOM_CERTIFICATE_FIND_APPLICATION_POLICY = 7,
  CAPICOM_CERTIFICATE_FIND_CERTIFICATE_POLICY = 8,
  CAPICOM_CERTIFICATE_FIND_TIME_VALID = 9,
  CAPICOM_CERTIFICATE_FIND_TIME_NOT_YET_VALID = 10,
  CAPICOM_CERTIFICATE_FIND_TIME_EXPIRED = 11,
  CAPICOM_CERTIFICATE_FIND_KEY_USAGE = 12,
}
// ...
export enum CAPICOM_KEY_USAGE {
  CAPICOM_DIGITAL_SIGNATURE_KEY_USAGE = 128,
  CAPICOM_NON_REPUDIATION_KEY_USAGE = 64,
  CAPICOM_KEY_ENCIPHERMENT_KEY_USAGE = 32,
  CAPICOM_DATA_ENCIPHERMENT_KEY_USAGE = 16,
  CAPICOM_KEY_AGREEMENT_KEY_USAGE = 8,
  CAPICOM_KEY_CERT_SIGN_KEY_USAGE = 4,
  CAPICOM_OFFLINE_CRL_SIGN_KEY_USAGE = 2,
  CAPICOM_CRL_SIGN_KEY_USAGE = 2,
  CAPICOM_ENCIPHER_ONLY_KEY_USAGE = 1,
  CAPICOM_DECIPHER_ONLY_KEY_USAGE = 32768,
}